requirements.txt
streamlit>=1.18.0
google-generativeai
chromadb
python-dotenv